Trying to handle the 401 errors better
I really need to figure out what is causing the 401 errors, and this is my attempt to do so. Adding in a retry to the API section, since the retweet seems to be where it fails with the 401 errors. Since this could add a delay into the reading of the stream, I wanted to throw in the console warnings about stalls from Twitter. I honestly have no idea if this will work.